requirements.txt
matplotlib
#tensorflow
#keras
imgaug
setuptools
keras_applications
pandas
scikit_optimize
click
numpy
opencv-contrib-python
pillow